requirements.txt
Django==2.2.2
Mastodon.py==1.4.5
praw==6.3.1
prawcore==1.0.1
